Rearranging code into a better structure: - Move tests project to tests/ from src/ - Extract out core business logic into netstandard20 project GildedRose.Core - Minor cleanups in sln file
