+++ src/GildedRose.Console/Program.cs
@@ -1,17 +1,48 @@
 ﻿using System.Collections.Generic;
+using GildedRose.Core;
 
 namespace GildedRose.Console
 {
     public class Program
     {
+        private readonly GildedRoseApp _app;
+
+        /// <summary>
+        /// Initializes the program / app with default inventory items.
+        /// </summary>
+        public Program() : this(new GildedRoseApp(GetDefaultItems())) { }
+
+        /// <summary>
+        /// Initializes the program with the given underlying app.
+        /// </summary>
+        /// <param name="app">Application containing business logic for GildedRose inventory management.</param>
+        public Program(GildedRoseApp app)
+        {
+            _app = app;
+        }
+
+        /// <summary>
+        /// Updates the quality and days-to-expiry (SellIn) values of the products in the inventory.
+        /// </summary>
+        // Note: Keeping `UpdateQuality()` method signature untouched as per the requirements.
+        public void UpdateQuality()
+        {
+            _app.UpdateQuality();
+        }
+
         static void Main(string[] args)
         {
-            System.Console.WriteLine("OMGHAI!");
-            var app = new GildedRoseApp(GetDefaultItems());
-            app.UpdateQuality();
+            System.Console.WriteLine("OMGHAI!"); // Not sure if this is needed.
+            var program = new Program();
+            program.UpdateQuality();
             System.Console.ReadKey();
         }
 
+        /// <summary>
+        /// Helper method to get default inventory items for initializing stock.
+        /// </summary>
+        /// <returns>List of default inventory items.</returns>
+        // TODO - Consider refactoring this so that the items are not hard-coded.
         private static IReadOnlyList<Item> GetDefaultItems() => new List<Item>
         {
             new Item {Name = "+5 Dexterity Vest", SellIn = 10, Quality = 20},
