--- src/GildedRose.Tests/TestAssemblyTests.cs
@@ -1,13 +1,0 @@
-using Xunit;
-
-namespace GildedRose.Tests
-{
-    public class TestAssemblyTests
-    {
-        [Fact]
-        public void TestTheTruth()
-        {
-            Assert.True(true);
-        }
-    }
-}
