+++ tests/GildedRose.Tests/TestAssemblyTests.cs
@@ -1,0 +1,13 @@
+using Xunit;
+
+namespace GildedRose.Tests
+{
+    public class TestAssemblyTests
+    {
+        [Fact]
+        public void TestTheTruth()
+        {
+            Assert.True(true);
+        }
+    }
+}
